# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Herdwise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List treatments
      parameters:
        - in: query
          name: cattleId
          schema:
            type: integer
        - in: query
          name: upcoming
          schema:
            type: boolean
      responses:
        "200":
          description: List of treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      summary: Record a treatment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/{id}:
    patch:
      operationId: updateTreatment
      tags: [treatments]
      summary: Update treatment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentUpdate"
      responses:
        "200":
          description: Updated treatment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTreatment
      tags: [treatments]
      summary: Delete treatment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cattle/{id}/weights:
    get:
      operationId: listWeightRecords
      tags: [cattle]
      summary: List weight history for an animal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of weight records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addWeightRecord
      tags: [cattle]
      summary: Add a weight record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cattle/{id}/weights/{weightId}:
    delete:
      operationId: deleteWeightRecord
      tags: [cattle]
      summary: Delete a weight record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: weightId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Treatment:
      type: object
      required: [id, cattleId, treatmentType, scheduledDate, completed, createdAt]
      properties:
        id:
          type: integer
        cattleId:
          type: integer
        cattleTag:
          type: ["string", "null"]
        treatmentType:
          type: string
        scheduledDate:
          type: string
        completedAt:
          type: ["string", "null"]
        completed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TreatmentInput:
      type: object
      required: [cattleId, treatmentType, scheduledDate]
      properties:
        cattleId:
          type: integer
        treatmentType:
          type: string
          minLength: 1
        scheduledDate:
          type: string
        notes:
          type: ["string", "null"]
    TreatmentUpdate:
      type: object
      properties:
        treatmentType:
          type: string
          minLength: 1
        scheduledDate:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    WeightRecord:
      type: object
      required: [id, cattleId, weightKg, recordedAt, createdAt]
      properties:
        id:
          type: integer
        cattleId:
          type: integer
        weightKg:
          type: number
        recordedAt:
          type: string
        notes:
          type: ["string", "null"]
        recordedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    WeightRecordInput:
      type: object
      required: [weightKg, recordedAt]
      properties:
        weightKg:
          type: number
          minimum: 0
        recordedAt:
          type: string
        notes:
          type: ["string", "null"]
        recordedBy:
          type: ["string", "null"]
...